Chapter05/restful_python_2_05/Django01/requirements.txt
Django==2.1.2
djangorestframework==3.9.0
httpie==0.9.9
